third_party/WebKit/LayoutTests/fast/forms/radio/radio-attr-order.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<input type="radio" checked name="test_group" id="rb">
<script>
description('This tests that radio buttons are checked by default, even when the name attribute is parsed after the checked attribute.');

shouldBeTrue('document.getElementById("rb").checked');
</script>
</body>
</html>
